! 2010-09-24 - Chris Jones  - Fix compilation warning on windows, due to implicit double to float conversion
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@86590 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: 86b23e1c9dc79841ae0edee833fecb037f8209a0

+++ LHCb/Event/PackedEvent/src/PackedWeightsVector.cpp
@@ -72,7 +72,7 @@ void WeightsVectorPacker::unpack( const PackedDataVector & pweightsV,
       {
         const PackedWeight & pweight = pweightsV.weights()[iW];
         wWeights.push_back( LHCb::WeightsVector::WeightData( pweight.key,
-                                                             m_pack.fraction(pweight.weight) ) );
+                                                             (float)m_pack.fraction(pweight.weight) ) );
       }
 
     }
